+++ components/DarkmodeToggle.tsx
@@ -1,0 +1,9 @@
+import { ModeToggle } from './ui/ModeToggle';
+
+export default function DarkmodeToggle() {
+  return (
+    <div className="fixed bottom-pr-60 right-pr-40 z-50">
+      <ModeToggle />
+    </div>
+  );
+}
